refactor(clawdstrike): extract policy tests into sibling tests/ module
Move the inline #[cfg(test)] mod tests (70 fns) out of policy/mod.rs into a
sibling child module policy/tests/ split into 5 theme files:
- schema_validation.rs (25): globs/regex/placeholder/version-gating + create_guards
- merge_extends.rs (18): merge strategies + extends inheritance + secret_leak merge
- spider_sense_merge.rs (8): spider_sense parse + deep-merge presence tracking
- origins.rs (9): origins parse/version-gating/merge
- rulesets_and_verifier.rs (10): built-in rulesets + load-time verifier

tests/mod.rs (declared #[cfg(test)] mod tests; from mod.rs) re-exports the shared
imports as pub(super) use so each theme file needs only . ENV_MUTEX +
Mutex live in schema_validation.rs (its only users). Every test body is byte-identical
(all 50 raw-string YAML literals verified unchanged); 1059/0/8 pass, same as baseline.
No test renamed, weakened, or ignored. mod.rs is now 63 lines.
